backend/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
identify==2.6.12
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pyflakes==3.3.2
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
virtualenv==20.31.2
